Refactor a little and grab all jobs
